Renamed Rack::File to Rack::Files, since it can serve multiple files from a root directory.
* Left a Rack::File constant, for backwards compatibility.
