+++ test/spec_files.rb
@@ -1,11 +1,11 @@
-require 'rack/file'
+require 'rack/files'
 require 'rack/mock'
 
-describe Rack::File do
+describe Rack::Files do
   DOCROOT = File.expand_path(File.dirname(__FILE__)) unless defined? DOCROOT
 
   should "serve files" do
-    res = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT))).
+    res = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT))).
       get("/cgi/test")
 
     res.should.be.ok
@@ -13,7 +13,7 @@ describe Rack::File do
   end
 
   should "set Last-Modified header" do
-    res = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT))).
+    res = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT))).
       get("/cgi/test")
 
     path = File.join(DOCROOT, "/cgi/test")
@@ -24,7 +24,7 @@ describe Rack::File do
 
   should "return 304 if file isn't modified since last serve" do
     path = File.join(DOCROOT, "/cgi/test")
-    res = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT))).
+    res = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT))).
       get("/cgi/test", 'HTTP_IF_MODIFIED_SINCE' => File.mtime(path).httpdate)
 
     res.status.should.equal 304
@@ -33,14 +33,14 @@ describe Rack::File do
 
   should "return the file if it's modified since last serve" do
     path = File.join(DOCROOT, "/cgi/test")
-    res = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT))).
+    res = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT))).
       get("/cgi/test", 'HTTP_IF_MODIFIED_SINCE' => (File.mtime(path) - 100).httpdate)
 
     res.should.be.ok
   end
 
   should "serve files with URL encoded filenames" do
-    res = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT))).
+    res = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT))).
       get("/cgi/%74%65%73%74") # "/cgi/test"
 
     res.should.be.ok
@@ -48,7 +48,7 @@ describe Rack::File do
   end
 
   should "allow safe directory traversal" do
-    req = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT)))
+    req = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT)))
 
     res = req.get('/cgi/../cgi/test')
     res.should.be.successful
@@ -61,7 +61,7 @@ describe Rack::File do
   end
 
   should "not allow unsafe directory traversal" do
-    req = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT)))
+    req = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT)))
 
     res = req.get("/../README")
     res.should.be.client_error
@@ -76,7 +76,7 @@ describe Rack::File do
   end
 
   should "allow files with .. in their name" do
-    req = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT)))
+    req = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT)))
     res = req.get("/cgi/..test")
     res.should.be.not_found
 
@@ -88,7 +88,7 @@ describe Rack::File do
   end
 
   should "not allow unsafe directory traversal with encoded periods" do
-    res = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT))).
+    res = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT))).
       get("/%2E%2E/README")
 
     res.should.be.client_error?
@@ -96,21 +96,21 @@ describe Rack::File do
   end
 
   should "allow safe directory traversal with encoded periods" do
-    res = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT))).
+    res = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT))).
       get("/cgi/%2E%2E/cgi/test")
 
     res.should.be.successful
   end
 
   should "404 if it can't find the file" do
-    res = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT))).
+    res = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT))).
       get("/cgi/blubb")
 
     res.should.be.not_found
   end
 
   should "detect SystemCallErrors" do
-    res = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT))).
+    res = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT))).
       get("/cgi")
 
     res.should.be.not_found
@@ -118,7 +118,7 @@ describe Rack::File do
 
   should "return bodies that respond to #to_path" do
     env = Rack::MockRequest.env_for("/cgi/test")
-    status, _, body = Rack::File.new(DOCROOT).call(env)
+    status, _, body = Rack::Files.new(DOCROOT).call(env)
 
     path = File.join(DOCROOT, "/cgi/test")
 
@@ -130,7 +130,7 @@ describe Rack::File do
   should "return correct byte range in body" do
     env = Rack::MockRequest.env_for("/cgi/test")
     env["HTTP_RANGE"] = "bytes=22-33"
-    res = Rack::MockResponse.new(*Rack::File.new(DOCROOT).call(env))
+    res = Rack::MockResponse.new(*Rack::Files.new(DOCROOT).call(env))
 
     res.status.should.equal 206
     res["Content-Length"].should.equal "12"
@@ -141,7 +141,7 @@ describe Rack::File do
   should "return error for unsatisfiable byte range" do
     env = Rack::MockRequest.env_for("/cgi/test")
     env["HTTP_RANGE"] = "bytes=1234-5678"
-    res = Rack::MockResponse.new(*Rack::File.new(DOCROOT).call(env))
+    res = Rack::MockResponse.new(*Rack::Files.new(DOCROOT).call(env))
 
     res.status.should.equal 416
     res["Content-Range"].should.equal "bytes */193"
@@ -149,14 +149,14 @@ describe Rack::File do
 
   should "support cache control options" do
     env = Rack::MockRequest.env_for("/cgi/test")
-    status, heads, _ = Rack::File.new(DOCROOT, 'public, max-age=38').call(env)
+    status, heads, _ = Rack::Files.new(DOCROOT, 'public, max-age=38').call(env)
 
     status.should.equal 200
     heads['Cache-Control'].should.equal 'public, max-age=38'
   end
 
   should "only support GET and HEAD requests" do
-    req = Rack::MockRequest.new(Rack::Lint.new(Rack::File.new(DOCROOT)))
+    req = Rack::MockRequest.new(Rack::Lint.new(Rack::Files.new(DOCROOT)))
 
     forbidden = %w[post put delete]
     forbidden.each do |method|
